+++ src/scielo/bin/PMCXMLTools/converter/pmcxml_files_set.py
@@ -1,6 +1,7 @@
 import shutil
 import os
 
+
 class PMCXML_FilesSet:
 
     def __init__(self, main_path, journal_folder, issue_folder, db_name):
@@ -10,10 +11,11 @@ class PMCXML_FilesSet:
         self.extracted_package_path = path + '/pmc_package_extracted' 
         self.db_path = path + '/pmc_base' 
         self.xml_path = path + '/pmc_xml'
+        self.jpg_path = path + '/img'
         #self.xml_path = path + '/xml' 
         #self.img_path = path + '/pmc_img' 
         #self.pdf_path = path + '/pmc_pdf' 
-        
+        self.img_converter = ImageConverter()
 
         self.db_name = db_name
 
@@ -23,11 +25,14 @@ class PMCXML_FilesSet:
         self.err_filename = self.db_path + '/' + db_name + '.err.log'
         
     def prepare_db_folder(self):
-        self.clean_folder(self.db_path, ['.id', '.mst', '.xrf', '.log', ])
+        self.delete_files_by_extension(self.db_path, ['.id', '.mst', '.xrf', '.log', ])
+    
+
+    def delete_db(self):
+        self.delete_files_by_extension(self.db_path, [ '.mst', '.xrf' ])
     
     def archive(self, xml_filename):
-        self.move_file_to_path(xml_filename, self.xml_path)
-        self.move_extracted_files_to_their_paths(xml_filename)
+        self.move_related_files(xml_filename)
 
         
 
@@ -50,34 +55,42 @@ class PMCXML_FilesSet:
 
     def move_file_to_path(self, filename, dest_path):
         f = os.path.basename(filename)
-        if not os.path.exists(dest_path):
-            os.makedirs(dest_path)
         if os.path.exists(dest_path + '/' + f):
             os.unlink(dest_path + '/' + f)
-        shutil.move(filename, dest_path)
+        if os.path.exists(filename):
+            
+            if not os.path.exists(dest_path):
+                os.makedirs(dest_path)
+    
+            shutil.move(filename, dest_path)
         return os.path.exists(dest_path + '/' + f)
     
-    def move_extracted_files_to_their_paths(self, xml_filename):
+    def move_related_files(self, xml_filename):
         count = 0
-        f = os.path.basename(xml_filename).replace('.xml', '') + '-'
-        path = os.path.dirname(xml_filename)
-        matched_files = [ filename for filename in os.listdir(path) if f in filename ]
-        #matched_files.append(os.path.basename(xml_filename))
-        for matched_file in matched_files:
+        matched_count = 0
+        for c in '-.':
+            pattern = os.path.basename(xml_filename).replace('.xml', '') + c
+            path = os.path.dirname(xml_filename)
+            matched_files = [ filename for filename in os.listdir(path) if pattern in filename ]
+            #matched_files.append(os.path.basename(xml_filename))
+            matched_count += len(matched_files)
+            for matched_file in matched_files:
             
-            if self.move_file_to_path(path + '/' + matched_file, self.xml_path):
-                count += 1
-        return (count == len(matched_files))
+                if self.move_file_to_path(path + '/' + matched_file, self.xml_path):
+                    count += 1
 
-    def clean_folder(self, path, array_extension):
+        return (count == matched_count)
+
+    def delete_files_by_extension(self, path, array_extension):
         if os.path.exists(path):
             files = os.listdir(path)
             for f in files:
-                print('deleting ' + path + '/' + f + '?')
+                #print('deleting ' + path + '/' + f + '?')
                 ext = f[f.rfind('.'):]
                 if ext in array_extension:
-                    print('deleting ' + path + '/' + f)
+                    #print('deleting ' + path + '/' + f)
                     os.remove(path + '/' + f)
         else:
             os.makedirs(path)
+
